// src/app/interfaces/messages.ts
// Generated by https://quicktype.io

export interface IResponseSendMessage {
  ok: boolean;
  message: string;
  data: IMessage;
}

export interface IMessage {
  id?: number;
  user_message: string | null;
  bot_response: string | null;
  language?: string;
  util?: boolean | null;
  create_at?: string;
  update_at?: string | null;
  temp_number?: number | null;
}